+++ ui/pages/Project/components/ProjectOverview.jsx
@@ -2,11 +2,12 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import sortBy from 'lodash/sortBy'
 import styled from 'styled-components'
-import { Grid, Icon, Popup } from 'semantic-ui-react'
+import { Grid, Icon, Popup, Loader, Dimmer } from 'semantic-ui-react'
 import { connect } from 'react-redux'
 import { NavLink } from 'react-router-dom'
 
 import { getUser } from 'redux/selectors'
+import DataLoader from 'shared/components/DataLoader'
 import { VerticalSpacer } from 'shared/components/Spacers'
 import UpdateButton from 'shared/components/buttons/UpdateButton'
 import { validators } from 'shared/components/form/ReduxFormWrapper'
@@ -20,14 +21,14 @@ import {
   DATASET_TYPE_SV_CALLS,
   ANVIL_URL,
 } from 'shared/utils/constants'
-import { updateProjectMmeContact } from '../reducers'
+import { updateProjectMmeContact, loadMmeSubmissions } from '../reducers'
 import {
   getAnalysisStatusCounts,
   getProjectAnalysisGroupFamiliesByGuid,
   getProjectAnalysisGroupIndividualsCount,
   getProjectAnalysisGroupSamplesByTypes,
-  getProjectAnalysisGroupMmeSubmissions,
   getProjectAnalysisGroupMmeSubmissionDetails,
+  getMmeSubmissionsLoading,
 } from '../selectors'
 import EditFamiliesAndIndividualsButton from './edit-families-and-individuals/EditFamiliesAndIndividualsButton'
 import EditIndividualMetadataButton from './edit-families-and-individuals/EditIndividualMetadataButton'
@@ -93,8 +94,8 @@ const MME_CONTACT_FIELDS = [
   },
 ]
 
-const BaseMatchmakerSubmissionOverview = React.memo(({ project, mmeSubmissions, onSubmit }) => (
-  <div>
+const BaseMatchmakerSubmissionOverview = React.memo(({ project, mmeSubmissions, onSubmit, load, loading }) => (
+  <DataLoader load={load} loading={false} content>
     {project.canEdit && (
       <UpdateButton
         onSubmit={onSubmit}
@@ -111,25 +112,30 @@ const BaseMatchmakerSubmissionOverview = React.memo(({ project, mmeSubmissions, 
     <DataTable
       basic="very"
       fixed
+      loading={loading}
       data={Object.values(mmeSubmissions)}
       idField="submissionGuid"
       defaultSortColumn="familyName"
       columns={MME_COLUMNS}
     />
-  </div>
+  </DataLoader>
 ))
 
 BaseMatchmakerSubmissionOverview.propTypes = {
   mmeSubmissions: PropTypes.arrayOf(PropTypes.object),
+  loading: PropTypes.bool,
+  load: PropTypes.func,
   project: PropTypes.object,
   onSubmit: PropTypes.func,
 }
 
 const mapMatchmakerSubmissionsStateToProps = (state, ownProps) => ({
   mmeSubmissions: getProjectAnalysisGroupMmeSubmissionDetails(state, ownProps),
+  loading: getMmeSubmissionsLoading(state),
 })
 
 const mapDispatchToProps = {
+  load: loadMmeSubmissions,
   onSubmit: updateProjectMmeContact,
 }
 
@@ -179,41 +185,29 @@ const mapFamiliesStateToProps = (state, ownProps) => ({
 
 const FamiliesIndividualsOverview = connect(mapFamiliesStateToProps)(FamiliesIndividuals)
 
-const Matchmaker = React.memo(({ project, mmeSubmissions }) => {
-  const mmeSubmissionCount = mmeSubmissions.length
-  const deletedSubmissionCount = mmeSubmissions.filter(({ deletedDate }) => deletedDate).length
+const MatchmakerOverview = React.memo(({ project }) => (
+  <DetailSection
+    title="Matchmaker Submissions"
+    content={project.mmeSubmissionCount ? (
+      <div>
+        {`${project.mmeSubmissionCount} submissions `}
+        <Modal
+          trigger={<ButtonLink icon="external" size="tiny" />}
+          title={`Matchmaker Submissions for ${project.name}`}
+          modalName="mmeSubmissions"
+          size="large"
+        >
+          <MatchmakerSubmissionOverview project={project} />
+        </Modal>
+        {project.mmeDeletedSubmissionCount > 0 && <div>{`${project.mmeDeletedSubmissionCount} removed submissions`}</div>}
+      </div>
+    ) : 'No Submissions'}
+  />
+))
 
-  return (
-    <DetailSection
-      title="Matchmaker Submissions"
-      content={mmeSubmissionCount ? (
-        <div>
-          {`${mmeSubmissionCount - deletedSubmissionCount} submissions `}
-          <Modal
-            trigger={<ButtonLink icon="external" size="tiny" />}
-            title={`Matchmaker Submissions for ${project.name}`}
-            modalName="mmeSubmissions"
-            size="large"
-          >
-            <MatchmakerSubmissionOverview project={project} />
-          </Modal>
-          {deletedSubmissionCount > 0 && <div>{`${deletedSubmissionCount} removed submissions`}</div>}
-        </div>
-      ) : 'No Submissions'}
-    />
-  )
-})
-
-Matchmaker.propTypes = {
+MatchmakerOverview.propTypes = {
   project: PropTypes.object.isRequired,
-  mmeSubmissions: PropTypes.arrayOf(PropTypes.object),
 }
-
-const mapMatchmakerStateToProps = (state, ownProps) => ({
-  mmeSubmissions: getProjectAnalysisGroupMmeSubmissions(state, ownProps),
-})
-
-const MatchmakerOverview = connect(mapMatchmakerStateToProps)(Matchmaker)
 
 class DatasetSection extends React.PureComponent {
 
@@ -351,10 +345,10 @@ const mapAnalysisStatusStateToProps = (state, ownProps) => ({
 
 const AnalysisStatusOverview = connect(mapAnalysisStatusStateToProps)(AnalysisStatus)
 
-const ProjectOverview = React.memo(props => (
+const ProjectOverview = React.memo(({ familiesLoading, ...props }) => (
   <Grid>
     <Grid.Column width={5}>
-      <FamiliesIndividualsOverview {...props} />
+      {familiesLoading ? <Dimmer inverted active><Loader /></Dimmer> : <FamiliesIndividualsOverview {...props} />}
       <VerticalSpacer height={10} />
       <MatchmakerOverview {...props} />
     </Grid.Column>
@@ -364,13 +358,14 @@ const ProjectOverview = React.memo(props => (
     </Grid.Column>
     <Grid.Column width={6}>
       <AnvilOverview {...props} />
-      <AnalysisStatusOverview {...props} />
+      {familiesLoading ? <Dimmer inverted active><Loader /></Dimmer> : <AnalysisStatusOverview {...props} />}
     </Grid.Column>
   </Grid>
 ))
 
 ProjectOverview.propTypes = {
   project: PropTypes.object.isRequired,
+  familiesLoading: PropTypes.bool,
 }
 
 export default ProjectOverview